# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PipeForge API
  version: 0.4.0-multipart
  description: >-
    Implemented HTTP contract for the Go control plane foundation, identity
    flows, and immutable dataset metadata with streamed uploads and presigned multipart sessions.
servers:
  - url: http://localhost:58080
paths:
...
  /v1/datasets/{datasetID}/uploads:
    post:
      operationId: initiateMultipartUpload
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: datasetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateMultipartUploadRequest'
      responses:
        '201':
          description: Multipart upload session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipartUploadSession'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/UploadTooLarge'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '502':
          $ref: '#/components/responses/StorageError'
  /v1/uploads/{sessionID}:
    get:
      operationId: getMultipartUpload
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Upload session and registered parts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipartUploadSessionDetail'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/uploads/{sessionID}/parts:
    get:
      operationId: listMultipartParts
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Registered multipart parts.
          content:
            application/json:
              schema:
                type: object
                required: [parts]
                properties:
                  parts:
                    type: array
                    items:
                      $ref: '#/components/schemas/MultipartUploadPart'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/uploads/{sessionID}/parts/{partNumber}/presign:
    post:
      operationId: presignMultipartPart
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: partNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Short-lived presigned URL for the part.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipartPartURL'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '502':
          $ref: '#/components/responses/StorageError'
  /v1/uploads/{sessionID}/parts/{partNumber}:
    put:
      operationId: registerMultipartPart
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: partNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterMultipartPartRequest'
      responses:
        '200':
          description: Part metadata registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipartUploadPart'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '502':
          $ref: '#/components/responses/StorageError'
  /v1/uploads/{sessionID}/complete:
    post:
      operationId: completeMultipartUpload
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMultipartUploadRequest'
      responses:
        '201':
          description: Immutable dataset version promoted to available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersion'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '413':
          $ref: '#/components/responses/UploadTooLarge'
        '502':
          $ref: '#/components/responses/StorageError'
  /v1/uploads/{sessionID}/abort:
    post:
      operationId: abortMultipartUpload
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Upload session aborted; repeated aborts are safe.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '502':
          $ref: '#/components/responses/StorageError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    InvalidRequest:
      description: The request is invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: Authentication is missing, invalid, expired, or revoked.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The authenticated principal is not allowed to access the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ResourceConflict:
      description: The resource cannot be changed in its current state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UploadTooLarge:
      description: The upload exceeds the configured maximum size.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    StorageError:
      description: Object storage could not complete or verify the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    DatasetVersion:
      type: object
      required: [id, datasetId, versionNumber, state, originalFilename, contentType, format, sourceInfo, artifactPrefix, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        datasetId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          format: int32
        state:
          type: string
          enum: [UPLOADING, AVAILABLE, FAILED]
        originalFilename:
          type: string
        contentType:
          type: string
        format:
          type: string
          enum: [CSV, JSONL, PARQUET]
        sizeBytes:
          type: integer
          format: int64
          nullable: true
        checksumSha256:
          type: string
          nullable: true
        rowCount:
          type: integer
          format: int64
          nullable: true
        columnCount:
          type: integer
          format: int32
          nullable: true
        encoding:
          type: string
          nullable: true
        compressionType:
          type: string
          nullable: true
        sourceInfo:
          type: object
          additionalProperties: true
        artifactPrefix:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        availableAt:
          type: string
          format: date-time
          nullable: true
...
    InitiateMultipartUploadRequest:
      type: object
      required: [filename, expectedSize]
      additionalProperties: false
      properties:
        filename:
          type: string
          maxLength: 255
        contentType:
          type: string
          default: application/octet-stream
        expectedSize:
          type: integer
          format: int64
          minimum: 1
        expectedChecksumSha256:
          type: string
          pattern: '^[0-9a-fA-F]{64}$'
        idempotencyKey:
          type: string
          maxLength: 128
    MultipartUploadSession:
      type: object
      required: [id, ownerUserId, datasetId, versionId, state, originalFilename, contentType, format, expectedSize, partSize, partCount, expiresAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        datasetId:
          type: string
          format: uuid
        versionId:
          type: string
          format: uuid
        state:
          type: string
          enum: [INITIATED, COMPLETING, ABORTING, COMPLETED, ABORTED, EXPIRED, FAILED]
        originalFilename:
          type: string
        contentType:
          type: string
        format:
          type: string
          enum: [CSV, JSONL, PARQUET]
        expectedSize:
          type: integer
          format: int64
        expectedChecksumSha256:
          type: string
          nullable: true
        partSize:
          type: integer
          format: int64
        partCount:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        abortedAt:
          type: string
          format: date-time
          nullable: true
    MultipartUploadPart:
      type: object
      required: [sessionId, partNumber, etag, sizeBytes, createdAt]
      properties:
        sessionId:
          type: string
          format: uuid
        partNumber:
          type: integer
          minimum: 1
        etag:
          type: string
        sizeBytes:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
    MultipartUploadSessionDetail:
      type: object
      required: [session, parts]
      properties:
        session:
          $ref: '#/components/schemas/MultipartUploadSession'
        parts:
          type: array
          items:
            $ref: '#/components/schemas/MultipartUploadPart'
    MultipartPartURL:
      type: object
      required: [partNumber, url, method, expiresAt]
      properties:
        partNumber:
          type: integer
        url:
          type: string
          format: uri
        method:
          type: string
          enum: [PUT]
        expiresAt:
          type: string
          format: date-time
    RegisterMultipartPartRequest:
      type: object
      required: [etag, sizeBytes]
      additionalProperties: false
      properties:
        etag:
          type: string
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
    CompleteMultipartUploadRequest:
      type: object
      required: [parts]
      additionalProperties: false
      properties:
        parts:
          type: array
          minItems: 1
          items:
            type: object
            required: [partNumber, etag]
            additionalProperties: false
            properties:
              partNumber:
                type: integer
                minimum: 1
              etag:
                type: string
...
    Problem:
      type: object
      required: [statusCode, code, message, requestId]
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true